requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
ruamel.yaml>=0.15
